requirements.txt
asgiref==3.7.2
Django==4.2.9
django-cors-headers==4.3.1
django-restframework==0.0.1
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2023.4